Address LSM write review comments
