+++ hudi-client/hudi-client-common/src/main/java/org/apache/hudi/io/HoodieAvroNativeCDCLogger.java
@@ -20,11 +20,8 @@ package org.apache.hudi.io;
 
 import org.apache.hudi.avro.HoodieAvroUtils;
 import org.apache.hudi.common.engine.TaskContextSupplier;
-import org.apache.hudi.common.fs.FSUtils;
 import org.apache.hudi.common.model.HoodieAvroIndexedRecord;
-import org.apache.hudi.common.model.HoodieFileFormat;
 import org.apache.hudi.common.model.HoodieKey;
-import org.apache.hudi.common.model.HoodieLogFile;
 import org.apache.hudi.common.model.HoodieRecord;
 import org.apache.hudi.common.schema.HoodieSchema;
 import org.apache.hudi.common.schema.HoodieSchemaCache;
@@ -35,13 +32,9 @@ import org.apache.hudi.common.table.cdc.HoodieCDCSupplementalLoggingMode;
 import org.apache.hudi.common.table.cdc.HoodieCDCUtils;
 import org.apache.hudi.common.table.log.LogFileCreationCallback;
 import org.apache.hudi.common.util.Option;
-import org.apache.hudi.common.util.StringUtils;
 import org.apache.hudi.config.HoodieWriteConfig;
 import org.apache.hudi.exception.HoodieException;
 import org.apache.hudi.exception.HoodieIOException;
-import org.apache.hudi.exception.HoodieUpsertException;
-import org.apache.hudi.io.storage.HoodieFileWriter;
-import org.apache.hudi.io.storage.HoodieFileWriterFactory;
 import org.apache.hudi.storage.HoodieStorage;
 import org.apache.hudi.storage.StoragePath;
 
@@ -50,11 +43,7 @@ import org.apache.avro.generic.GenericRecord;
 import org.apache.avro.generic.IndexedRecord;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.Properties;
 
 import static org.apache.hudi.common.table.cdc.HoodieCDCSupplementalLoggingMode.DATA_BEFORE;
 import static org.apache.hudi.common.table.cdc.HoodieCDCSupplementalLoggingMode.DATA_BEFORE_AFTER;
@@ -66,22 +55,11 @@ public class HoodieAvroNativeCDCLogger implements HoodieCDCLogWriter<IndexedReco
 
   private final String commitTime;
   private final String partitionPath;
-  private final HoodieStorage storage;
-  private final HoodieWriteConfig config;
   private final HoodieSchema dataSchema;
   private final HoodieSchema cdcSchema;
-  private final HoodieFileFormat nativeFileFormat;
   private final HoodieCDCSupplementalLoggingMode cdcSupplementalLoggingMode;
-  private final StoragePath parentPath;
-  private final String fileId;
-  private final String writeToken;
-  private final LogFileCreationCallback fileCreationCallback;
-  private final TaskContextSupplier taskContextSupplier;
   private final CDCTransformer transformer;
-  private final Properties recordProperties;
-  private final List<StoragePath> cdcAbsPaths;
-  private int nextLogVersion;
-  private HoodieFileWriter cdcWriter;
+  private final HoodieNativeCDCFileWriter nativeCDCFileWriter;
   private PendingCDCRecord pendingRecord;
 
   public HoodieAvroNativeCDCLogger(
@@ -98,22 +76,23 @@ public class HoodieAvroNativeCDCLogger implements HoodieCDCLogWriter<IndexedReco
       TaskContextSupplier taskContextSupplier) {
     this.commitTime = commitTime;
     this.partitionPath = partitionPath;
-    this.storage = storage;
-    this.config = config;
     this.dataSchema = HoodieSchemaCache.intern(HoodieSchemaUtils.removeMetadataFields(schema));
     this.cdcSupplementalLoggingMode = tableConfig.cdcSupplementalLoggingMode();
     this.cdcSchema = HoodieCDCUtils.schemaBySupplementalLoggingMode(cdcSupplementalLoggingMode, dataSchema);
-    this.nativeFileFormat = tableConfig.getBaseFileFormat();
-    this.parentPath = parentPath;
-    this.fileId = fileId;
-    this.writeToken = writeToken;
-    this.fileCreationCallback = fileCreationCallback;
-    this.taskContextSupplier = taskContextSupplier;
     this.transformer = getTransformer();
-    this.recordProperties = new Properties();
-    this.recordProperties.putAll(config.getProps());
-    this.cdcAbsPaths = new ArrayList<>();
-    this.nextLogVersion = HoodieLogFile.LOGFILE_BASE_VERSION;
+    this.nativeCDCFileWriter = new HoodieNativeCDCFileWriter(
+        commitTime,
+        partitionPath,
+        storage,
+        config,
+        cdcSchema,
+        tableConfig.getBaseFileFormat(),
+        parentPath,
+        fileId,
+        writeToken,
+        fileCreationCallback,
+        taskContextSupplier,
+        HoodieRecord.HoodieRecordType.AVRO);
   }
 
   @Override
@@ -142,24 +121,14 @@ public class HoodieAvroNativeCDCLogger implements HoodieCDCLogWriter<IndexedReco
 
   @Override
   public Map<String, Long> getCDCWriteStats() {
-    Map<String, Long> stats = new HashMap<>();
-    try {
-      for (StoragePath cdcAbsPath : cdcAbsPaths) {
-        String cdcFileName = cdcAbsPath.getName();
-        String cdcPath = StringUtils.isNullOrEmpty(partitionPath) ? cdcFileName : partitionPath + "/" + cdcFileName;
-        stats.put(cdcPath, storage.getPathInfo(cdcAbsPath).getLength());
-      }
-    } catch (IOException e) {
-      throw new HoodieUpsertException("Failed to get cdc write stat", e);
-    }
-    return stats;
+    return nativeCDCFileWriter.getCDCWriteStats();
   }
 
   @Override
   public void close() {
     try {
       flushPendingRecord();
-      closeCDCWriter();
+      nativeCDCFileWriter.close();
     } catch (IOException e) {
       throw new HoodieIOException("Failed to close HoodieAvroNativeCDCLogger", e);
     }
@@ -170,55 +139,13 @@ public class HoodieAvroNativeCDCLogger implements HoodieCDCLogWriter<IndexedReco
       return;
     }
     try {
-      ensureCDCWriter();
-      cdcWriter.write(
+      nativeCDCFileWriter.write(
           pendingRecord.recordKey,
-          new HoodieAvroIndexedRecord(new HoodieKey(pendingRecord.recordKey, partitionPath), pendingRecord.record),
-          cdcSchema,
-          recordProperties);
+          new HoodieAvroIndexedRecord(new HoodieKey(pendingRecord.recordKey, partitionPath), pendingRecord.record));
       pendingRecord = null;
     } catch (IOException e) {
       throw new HoodieException("Failed to write the cdc data to native cdc log file", e);
     }
-  }
-
-  private void ensureCDCWriter() throws IOException {
-    if (cdcWriter != null && cdcWriter.canWrite()) {
-      return;
-    }
-    closeCDCWriter();
-    HoodieLogFile cdcLogFile = createNativeCDCLogFile();
-    cdcWriter = HoodieFileWriterFactory.getFileWriter(
-        commitTime, cdcLogFile.getPath(), storage, config, cdcSchema, taskContextSupplier, HoodieRecord.HoodieRecordType.AVRO);
-    cdcAbsPaths.add(cdcLogFile.getPath());
-  }
-
-  private void closeCDCWriter() throws IOException {
-    if (cdcWriter != null) {
-      cdcWriter.close();
-      cdcWriter = null;
-    }
-  }
-
-  private HoodieLogFile createNativeCDCLogFile() throws IOException {
-    int version = nextAvailableVersion();
-    HoodieLogFile nativeCDCLogFile = new HoodieLogFile(makeNativeCDCLogPath(version), 0);
-    fileCreationCallback.preFileCreation(nativeCDCLogFile);
-    nextLogVersion = version + 1;
-    return nativeCDCLogFile;
-  }
-
-  private int nextAvailableVersion() throws IOException {
-    int candidateVersion = nextLogVersion;
-    while (storage.exists(makeNativeCDCLogPath(candidateVersion))) {
-      candidateVersion++;
-    }
-    return candidateVersion;
-  }
-
-  private StoragePath makeNativeCDCLogPath(int version) {
-    return new StoragePath(parentPath, FSUtils.makeNativeLogFileName(
-        fileId, writeToken, commitTime, version, HoodieCDCUtils.CDC_LOGFILE_SUFFIX, nativeFileFormat));
   }
 
   private CDCTransformer getTransformer() {
